# Optimize `method_missing` via `alias_method`